Moved Xml-based configuration to a new dll (Closes #84) - New dll StructureMap.Configuration.Xml   Once referenced, the user can call the same methods on the configuration   expression, just that they are now extension methods - All tests related to the xml config functionality has been moved   to Structuremap.Xml.Testing: The structure, though, has been flattened   and some parts of the DataMother have been duplicated, while some files - StructureMap itself does not know about xml-based config, which is   proven by the removed dependency to System.Xml
What is missing in this commit is setting up e.g. a Nuget that would bring
XmlConfig capabilities to StructureMap.
