+++ src/StructureMap.Configuration.Xml/StructureMapConfigurationSection.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using System.Configuration;
 using System.Xml;
 
-namespace StructureMap.Configuration
+namespace StructureMap.Configuration.Xml
 {
     public class StructureMapConfigurationSection : IConfigurationSectionHandler
     {
