+++ src/StructureMap.Testing/Graph/SetterInjectionTester.cs
@@ -19,49 +19,6 @@ namespace StructureMap.Testing.Graph
 
         #endregion
 
-        private PluginGraph getPluginGraph()
-        {
-            return
-                DataMother.BuildPluginGraphFromXml(
-                    @"
-<StructureMap>
-	<Assembly Name='StructureMap.Testing.Widget'/>
-	<Assembly Name='StructureMap.Testing.Widget5'/>
-	
-	<PluginFamily Type='StructureMap.Testing.Widget.IWidget' Assembly='StructureMap.Testing.Widget' DefaultKey='Money'>
-		<Source Type='XmlFile' FilePath='FullTesting.XML' XPath='Widgets' NodeName='Widget'/>
-		<Plugin Assembly='StructureMap.Testing.Widget' Type='StructureMap.Testing.Widget.NotPluggableWidget' ConcreteKey='NotPluggable'/>
-	</PluginFamily>
-
-	<PluginFamily Type='StructureMap.Testing.Widget.Rule' Assembly='StructureMap.Testing.Widget' DefaultKey='Blue'>
-		<Instance Key='Bigger' Type='GreaterThan'>
-			<Property Name='Attribute' Value='MyDad' />
-			<Property Name='Value' Value='10' />
-		</Instance>
-		<Instance Key='Blue' Type='Color'>
-			<Property Name='Color' Value='Blue' />
-		</Instance>
-		<Instance Key='Red' Type='Color'>
-			<Property Name='Color' Value='Red' />
-		</Instance>
-	</PluginFamily>
-
-	<PluginFamily Type='StructureMap.Testing.Widget5.IGridColumn' Assembly='StructureMap.Testing.Widget5' DefaultKey=''>
-		<Source Type='XmlFile' FilePath='GridColumnInstances.xml' XPath='//GridColumns' NodeName='GridColumn'/>
-		<Plugin Assembly='StructureMap.Testing.Widget5' Type='StructureMap.Testing.Widget5.OtherGridColumn' ConcreteKey='Other'>
-			<Setter Name='ColumnName' />
-			<Setter Name='FontStyle' />
-			<Setter Name='Rules' />
-			<Setter Name='Widget' />
-			<Setter Name='WrapLines' />
-		</Plugin>
-	</PluginFamily>
-	
-	<Instances/>
-</StructureMap>
-");
-        }
-
         public class SetterTarget
         {
             public string Name1 { get; set; }
