+++ src/StructureMap.Testing/StructureMap.Testing.csproj
@@ -204,12 +204,10 @@
     <Compile Include="Bugs\FillAllPropertiesShouldWorkForAlreadyConfiguredPluginsBug.cs" />
     <Compile Include="Bugs\GetAllInstancesShouldObserveKeys.cs" />
     <Compile Include="Bugs\HttpSessionNullRefBug.cs" />
-    <Compile Include="Bugs\IDictionaryAndXmlBugTester.cs" />
     <Compile Include="Bugs\InjectByFuncWithNoPublicConstructors.cs" />
     <Compile Include="Bugs\LambdaCreatesNullBugTester.cs" />
     <Compile Include="Bugs\MixedConfigureAndInitializeMissingInstanceProblem.cs" />
     <Compile Include="Bugs\OpenGenericWithConstraints.cs" />
-    <Compile Include="Bugs\OverwritingOfPrimitiveProperties.cs" />
     <Compile Include="Bugs\ScanIndexerBugTester.cs" />
     <Compile Include="Bugs\SingletonShouldBeLazy.cs" />
     <Compile Include="Bugs\SpecifyScopeInConfigureTester.cs" />
@@ -221,9 +219,6 @@
     <Compile Include="BuildUpIntegratedTester.cs" />
     <Compile Include="BuildUpTester.cs" />
     <Compile Include="ConcreteClassCreationTester.cs" />
-    <Compile Include="Configuration\DefaultInstanceNodeTester.cs" />
-    <Compile Include="Configuration\DictionaryAndArrayArgumentTester.cs" />
-    <Compile Include="Configuration\DictionaryReaderTester.cs" />
     <Compile Include="Configuration\DSL\AddInstanceTester.cs" />
     <Compile Include="Configuration\DSL\AddTypesTester.cs" />
     <Compile Include="Configuration\DSL\Add_PluginGraphConfigurationTester.cs" />
@@ -239,8 +234,6 @@
     <Compile Include="Configuration\DSL\RegistryIntegratedTester.cs" />
     <Compile Include="Configuration\DSL\RegistryTester.cs" />
     <Compile Include="Configuration\GraphBuilderTester.cs" />
-    <Compile Include="Configuration\PrimitiveArrayReaderTester.cs" />
-    <Compile Include="Configuration\ShortcuttedInstanceNodeTester.cs" />
     <Compile Include="Configuration\SimplePluginFactory.cs" />
     <Compile Include="Construction\BuilderCompilerTester.cs" />
     <Compile Include="Construction\ConstructorFunctionBuilderTester.cs" />
@@ -267,9 +260,6 @@
     <Compile Include="Graph\EnumerationTester.cs">
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="Graph\ExceptionHandling\StructureMapExceptionTester.cs">
-      <SubType>Code</SubType>
-    </Compile>
     <Compile Include="Graph\FindAllTypesFilterTester.cs" />
     <Compile Include="Graph\FirstInterfaceConventionTester.cs" />
     <Compile Include="Graph\GenericConnectionScannerTester.cs" />
@@ -286,9 +276,6 @@
     <Compile Include="Graph\Interceptors\InterceptorLibraryTester.cs" />
     <Compile Include="Graph\Interceptors\MockTypeInterceptor.cs" />
     <Compile Include="Graph\Interceptors\TypeInterceptionTester.cs" />
-    <Compile Include="Graph\Source\XmlAttributeInstanceMementoTester.cs">
-      <SubType>Code</SubType>
-    </Compile>
     <Compile Include="Graph\TypeFindingTester.cs" />
     <Compile Include="Diagnostics\TextReportWriterSmokeTester.cs" />
     <Compile Include="GenericsAcceptanceTester.cs" />
@@ -362,7 +349,6 @@
     <Compile Include="RegistryExtensions.cs" />
     <Compile Include="SessionCacheTester.cs" />
     <Compile Include="SpecificationExtensions.cs" />
-    <Compile Include="StructureMapConfigCreator.cs" />
     <Compile Include="StructureMapConfigurationTester.cs" />
     <Compile Include="TestData\DataMother.cs">
       <SubType>Code</SubType>
@@ -372,9 +358,6 @@
     <Compile Include="Diagnostics\ValidationBuildSessionTester.cs" />
     <Compile Include="TypeExtensions.cs" />
     <Compile Include="TypeExtensionsTester.cs" />
-    <Compile Include="XmlWriting\ElementChecker.cs">
-      <SubType>Code</SubType>
-    </Compile>
     <None Include="..\structuremap.snk">
       <Link>structuremap.snk</Link>
     </None>
@@ -384,9 +367,6 @@
       <LastGenOutput>Settings.Designer.cs</LastGenOutput>
     </None>
     <None Include="StructureMap.config">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-    <None Include="StructureMap.Testing.dll.config">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
     <EmbeddedResource Include="Graph\Source\EmbeddedMementoFile.xml" />
@@ -423,6 +403,9 @@
     <EmbeddedResource Include="TestData\Trigger.xml" />
     <EmbeddedResource Include="TestData\Valid.config" />
     <EmbeddedResource Include="TestData\XmlInstanceMemento.xml" />
+    <None Include="StructureMap.Testing.dll.config">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="TestData\GenericsTesting.xml" />
@@ -437,8 +420,6 @@
   <ItemGroup>
     <Content Include="Examples\AddRegistryInXml.xml" />
     <Content Include="Sample.xml" />
-    <EmbeddedResource Include="TestData\BUG_OverwritingOfPrimitiveProperties.xml" />
-    <EmbeddedResource Include="TestData\ShortInstance.xml" />
     <EmbeddedResource Include="TestData\DefaultInstance.xml" />
   </ItemGroup>
   <ItemGroup>
@@ -470,6 +451,7 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="AutoFactories\Supporting\" />
+    <Folder Include="Graph\ExceptionHandling\" />
     <Folder Include="Util\" />
   </ItemGroup>
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
