+++ src/StructureMap.Testing/StructureMap.Testing.dll.config
@@ -1,31 +1,5 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <configuration>
-  <configSections>
-    <section name="StructureMap" type="StructureMap.Configuration.StructureMapConfigurationSection,StructureMap"/>
-  </configSections>
-
-  <StructureMap>
-    <!-- Put StructureMap configuration here -->
-
-    <DefaultInstance 
-      PluginType="StructureMap.Testing.GenericWidgets.IThing`2, StructureMap.Testing.GenericWidgets" 
-      PluggedType="StructureMap.Testing.GenericWidgets.ColorThing`2, StructureMap.Testing.GenericWidgets" 
-      Key="Cornflower" 
-      color="Cornflower"/>
-    
-    <AddInstance 
-      Key="Complicated" 
-      PluginType="StructureMap.Testing.GenericWidgets.IThing`2, StructureMap.Testing.GenericWidgets" 
-      PluggedType="StructureMap.Testing.GenericWidgets.ComplexThing`2, StructureMap.Testing.GenericWidgets" 
-      name="Jeremy" 
-      age="32" 
-      ready="true"/>
-
-    <PluginFamily Assembly="StructureMap.Testing.GenericWidgets" Type="StructureMap.Testing.GenericWidgets.ISimpleThing`1" DefaultKey="Simple">
-      <Plugin Assembly="StructureMap.Testing.GenericWidgets" Type="StructureMap.Testing.GenericWidgets.SimpleThing`1" ConcreteKey="Simple" />
-    </PluginFamily>
-  </StructureMap>
-  
   
   <appSettings>
     <add key="Color" value="Blue"/>
