+++ src/StructureMap.Xml.Testing/Bugs/IFooWithPrimitives.cs
@@ -1,0 +1,8 @@
+﻿namespace StructureMap.Xml.Testing.Bugs
+{
+    public interface IFooWithPrimitives
+    {
+        bool IsTest { get; }
+        string TestValue { get; }
+    }
+}
