+++ src/StructureMap.Xml.Testing/StructureMap.Xml.Testing.dll.config
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+  <configSections>
+    <section name="StructureMap" type="StructureMap.Configuration.Xml.StructureMapConfigurationSection,StructureMap.Configuration.Xml"/>
+  </configSections>
+
+  <StructureMap>
+    <!-- Put StructureMap configuration here -->
+
+    <DefaultInstance 
+      PluginType="StructureMap.Testing.GenericWidgets.IThing`2, StructureMap.Testing.GenericWidgets" 
+      PluggedType="StructureMap.Testing.GenericWidgets.ColorThing`2, StructureMap.Testing.GenericWidgets" 
+      Key="Cornflower" 
+      color="Cornflower"/>
+    
+    <AddInstance 
+      Key="Complicated" 
+      PluginType="StructureMap.Testing.GenericWidgets.IThing`2, StructureMap.Testing.GenericWidgets" 
+      PluggedType="StructureMap.Testing.GenericWidgets.ComplexThing`2, StructureMap.Testing.GenericWidgets" 
+      name="Jeremy" 
+      age="32" 
+      ready="true"/>
+
+    <PluginFamily Assembly="StructureMap.Testing.GenericWidgets" Type="StructureMap.Testing.GenericWidgets.ISimpleThing`1" DefaultKey="Simple">
+      <Plugin Assembly="StructureMap.Testing.GenericWidgets" Type="StructureMap.Testing.GenericWidgets.SimpleThing`1" ConcreteKey="Simple" />
+    </PluginFamily>
+  </StructureMap>
+  
+</configuration>
