+++ src/StructureMap.Xml.Testing/TestData/BUG_OverwritingOfPrimitiveProperties.xml
@@ -1,0 +1,12 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<StructureMap>
+
+  <DefaultInstance 
+    PluginType="StructureMap.Xml.Testing.Bugs.IFooWithPrimitives, StructureMap.Xml.Testing"
+    PluggedType="StructureMap.Xml.Testing.Bugs.FooWithPrimitives, StructureMap.Xml.Testing"
+    Key="TEST"
+    name="Test123"
+    />
+  
+
+</StructureMap>
