+++ src/StructureMap.Xml.Testing/TestData/ShortInstance.xml
