+++ src/StructureMap/Configuration/ITypeReader.cs
@@ -4,9 +4,9 @@ using StructureMap.Pipeline;
 
 namespace StructureMap.Configuration
 {
-    public interface ITypeReader
+    public interface ITypeReader<T>
     {
         bool CanProcess(Type pluginType);
-        Instance Read(XmlNode node, Type pluginType);
+        Instance Read(T node, Type pluginType);
     }
 }
