+++ src/StructureMap/ConfigurationExpression.cs
@@ -1,5 +1,4 @@
 using System.Collections.Generic;
-using System.Xml;
 using StructureMap.Configuration;
 using StructureMap.Configuration.DSL;
 using StructureMap.Graph;
@@ -22,14 +21,6 @@ namespace StructureMap
         }
 
         /// <summary>
-        ///     If called, directs StructureMap to look for configuration in the App.config in any &lt;StructureMap&gt; node.
-        /// </summary>
-        public void IncludeConfigurationFromConfigFile()
-        {
-            ConfigurationParser.FromApplicationConfig().Each(x => _builder.Add(x));
-        }
-
-        /// <summary>
         ///     Creates and adds a Registry object of type T.
         /// </summary>
         /// <typeparam name="T">The Registry Type</typeparam>
@@ -45,27 +36,6 @@ namespace StructureMap
         public void AddRegistry(Registry registry)
         {
             _builder.Add(registry);
-        }
-
-        /// <summary>
-        ///     Imports configuration from an Xml file.  The fileName
-        ///     must point to an Xml file with valid StructureMap
-        ///     configuration
-        /// </summary>
-        public void AddConfigurationFromXmlFile(string fileName)
-        {
-            _builder.Add(ConfigurationParser.FromFile(fileName));
-        }
-
-        /// <summary>
-        ///     Imports configuration directly from an XmlNode.  This
-        ///     method was intended for scenarios like Xml being embedded
-        ///     into an assembly.  The node must be a 'StructureMap' node
-        /// </summary>
-        /// <param name="node"></param>
-        public void AddConfigurationFromNode(XmlNode node)
-        {
-            _builder.Add(new ConfigurationParser(node));
         }
 
         internal PluginGraph BuildGraph()
